refactor EventHelpers calendar code to helpers/NPCalendar
